SER-3478 switch close wiki to GCS
